fix: update components to ts, add docstrings, fix testing
