+++ src/js/sequences/components/Barcode/Add.jsx
@@ -8,7 +8,7 @@ import PersistForm from "../../../forms/components/PersistForm";
 import { addSequence } from "../../../otus/actions";
 import { getActiveIsolateId, getOTUDetailId } from "../../../otus/selectors";
 import { getDefaultTargetName, getUnreferencedTargets } from "../../selectors";
-import { SequenceForm, validationSchema } from "../Form";
+import { SequenceForm, validationSchema } from "../SequenceForm";
 import TargetsField from "./TargetField";
 
 const getInitialValues = defaultTarget => ({
