+++ src/js/sequences/components/Genome/__tests__/AddGenomeSequence.test.tsx
@@ -4,7 +4,7 @@ import userEvent from "@testing-library/user-event";
 import React from "react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { renderWithProviders } from "../../../../../tests/setupTests";
-import { castValues, EditGenomeSequence } from "../Edit";
+import { AddGenomeSequence, castValues } from "../AddGenomeSequence";
 
 function createAppStore(state) {
     return () =>
@@ -14,7 +14,7 @@ function createAppStore(state) {
         });
 }
 
-describe("<EditGenomeSequence>", () => {
+describe("<AddGenomeSequence>", () => {
     let props;
     let state;
 
@@ -22,16 +22,10 @@ describe("<EditGenomeSequence>", () => {
         props = {
             isolateId: "test_isolate_id",
             otuId: "test_otu_id",
-            id: "test_id",
             show: true,
             segments: [],
             onHide: vi.fn(),
             onSave: vi.fn(),
-            initialAccession: "initialAccession",
-            initialDefinition: "initialDefinition",
-            initialHost: "initialHost",
-            initialSegment: "test_segment",
-            initialSequence: "ATAG",
         };
         state = {
             otus: {
@@ -62,59 +56,37 @@ describe("<EditGenomeSequence>", () => {
         };
     });
 
-    it("should render all fields with current sequence data", () => {
-        renderWithProviders(<EditGenomeSequence {...props} />, createAppStore(state));
+    it("should update fields on typing", async () => {
+        renderWithProviders(<AddGenomeSequence {...props} />, createAppStore(state));
 
         expect(screen.getByText("Segment")).toBeInTheDocument();
         expect(screen.getByRole("combobox")).toBeInTheDocument();
-        expect(screen.getByRole("textbox", { name: "Accession (ID)" })).toHaveValue(props.initialAccession);
-        expect(screen.getByRole("textbox", { name: "Host" })).toHaveValue(props.initialHost);
-        expect(screen.getByRole("textbox", { name: "Definition" })).toHaveValue(props.initialDefinition);
-        expect(screen.getByRole("textbox", { name: "Sequence 4" })).toHaveValue(props.initialSequence);
-    });
+        expect(screen.getByRole("textbox", { name: "Accession (ID)" })).toBeInTheDocument();
+        expect(screen.getByRole("textbox", { name: "Host" })).toBeInTheDocument();
+        expect(screen.getByRole("textbox", { name: "Definition" })).toBeInTheDocument();
+        expect(screen.getByRole("textbox", { name: "Sequence 0" })).toBeInTheDocument();
 
-    it("should submit correct data when all fields changed", async () => {
-        renderWithProviders(<EditGenomeSequence {...props} />, createAppStore(state));
-
-        expect(screen.getByRole("combobox")).toBeInTheDocument();
-        await userEvent.click(screen.getByText("None"));
-
-        const accessionField = screen.getByRole("textbox", { name: "Accession (ID)" });
-        await userEvent.clear(accessionField);
-        await userEvent.type(accessionField, "user_typed_accession");
-
-        const hostField = screen.getByRole("textbox", { name: "Host" });
-        await userEvent.clear(hostField);
-        await userEvent.type(hostField, "user_typed_host");
-
-        const definitionField = screen.getByRole("textbox", { name: "Definition" });
-        await userEvent.clear(definitionField);
-        await userEvent.type(definitionField, "user_typed_definition");
-
-        const sequenceField = screen.getByRole("textbox", { name: "Sequence 4" });
-        await userEvent.clear(sequenceField);
-        await userEvent.type(sequenceField, "ACGRYKM");
-
+        await userEvent.click(screen.getByRole("combobox"));
+        await userEvent.click(screen.getByText("test_segment"));
+        await userEvent.type(screen.getByRole("textbox", { name: "Accession (ID)" }), "user_typed_accession");
+        await userEvent.type(screen.getByRole("textbox", { name: "Host" }), "user_typed_host");
+        await userEvent.type(screen.getByRole("textbox", { name: "Definition" }), "user_typed_definition");
+        await userEvent.type(screen.getByRole("textbox", { name: "Sequence 0" }), "ATGRYKM");
         await userEvent.click(screen.getByRole("button", { name: "Save" }));
 
         expect(props.onSave).toHaveBeenCalledWith(
             "test_otu_id",
             "test_isolate_id",
-            "test_id",
             "user_typed_accession",
             "user_typed_definition",
             "user_typed_host",
-            null,
-            "ACGRYKM",
+            "test_segment",
+            "ATGRYKM",
         );
     });
-
     it("should display errors when accession, definition, or sequence not defined", async () => {
-        renderWithProviders(<EditGenomeSequence {...props} />, createAppStore(state));
+        renderWithProviders(<AddGenomeSequence {...props} />, createAppStore(state));
 
-        await userEvent.clear(screen.getByRole("textbox", { name: "Accession (ID)" }));
-        await userEvent.clear(screen.getByRole("textbox", { name: "Definition" }));
-        await userEvent.clear(screen.getByRole("textbox", { name: "Sequence 4" }));
         await userEvent.click(screen.getByRole("button", { name: "Save" }));
 
         expect(props.onSave).not.toHaveBeenCalled();
@@ -125,23 +97,25 @@ describe("<EditGenomeSequence>", () => {
     });
 
     it("should display specific error when sequence contains chars !== ATCGNRYKM", async () => {
-        renderWithProviders(<EditGenomeSequence {...props} />, createAppStore(state));
+        renderWithProviders(<AddGenomeSequence {...props} />, createAppStore(state));
 
-        await userEvent.type(screen.getByRole("textbox", { name: "Sequence 4" }), "q");
+        const field = screen.getByRole("textbox", { name: /Sequence/ });
+
+        await userEvent.type(field, "atbcq");
         await userEvent.click(screen.getByRole("button", { name: "Save" }));
 
-        expect(screen.getByRole("textbox", { name: "Sequence 5" })).toHaveStyle("border: 1px solid #E0282E");
+        expect(field).toHaveStyle("border: 1px solid #E0282E");
         expect(screen.getByText("Sequence should only contain the characters: ATCGNRYKM")).toBeInTheDocument();
     });
 });
 
 describe("castValues", () => {
     const segments = [
-        { name: "test_1", molecule: "", required: true },
-        { name: "test_2", molecule: "", required: true },
+        { name: "test_1", molecule: null, required: true },
+        { name: "test_2", molecule: null, required: true },
     ];
 
-    const values = { segment: "test_1", otherData: {} };
+    const values = { segment: "test_1", sequence: "", accession: "", definition: "", host: "" };
 
     it("should return unchanged values when segment in selectable segments", () => {
         const castedValues = castValues(segments)(values);
@@ -151,6 +125,6 @@ describe("castValues", () => {
     it("should return values where segment: null when segment is not selectable", () => {
         values.segment = "invalid_segment";
         const castedValues = castValues(segments)(values);
-        expect(castedValues).toEqual({ ...values, segment: undefined });
+        expect(castedValues).toEqual({ ...values, segment: "" });
     });
 });
